requirements.txt
streamlit==1.28.0
mysql-connector-python==8.2.0
paramiko==3.4.0
pandas==2.2.0
plotly==5.17.0